Tweak the text on the three tenets
This uses what I propose is both punchier and more accurate language for
the three tenets of open companies. We don't share *everything*
(database passwords? all user data?). We "Share AMAP [as much as
possible]." Then we "Charge ALAP" to be parallel.

"Employ everyone" I think has a really nice ring to it!